Rename and reorder store names in config router
