+++ html/pfappserver/root/static.alt/src/views/Configuration/_store/floatingDevices.js
